backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
alembic==1.13.0
redis==5.1.0
celery==5.4.0
pinecone-client==3.2.0
sentence-transformers==3.0.0
openai==1.40.0
boto3==1.34.0
cloudinary==1.40.0
pyjwt==2.9.0
bcrypt==4.2.0
pydantic-settings==2.5.0
python-socketio==5.11.0
python-multipart==0.0.9
httpx==0.27.0
passlib==1.7.4
